pyproject.toml
[tool.poetry]
name = "albert-home-challenge"
version = "0.1.0"
description = ""
authors = ["Antoine Guidis <antoine@spot.so>"]
readme = "README.md"
packages = [{include = "albert_home_challenge"}]

[tool.taskipy.tasks]
api = "gunicorn --bind 0.0.0.0:8080 src.app:app --threads 2 --reload --access-logfile - --log-level debug"
lint = "ruff check src"
format = "black src"
test = "pytest -v"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.1.1"
gunicorn = "^23.0.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.12.2"
mypy = "^1.16.1"
taskipy = "^1.14.1"
python-dotenv = "^1.1.1"
pytest = "^8.4.1"
marshmallow = "^4.0.0"
webargs = "^8.7.0"

[tool.poetry.plugins."poetry.plugin"]
"taskipy" = "taskipy.plugin:TaskLoader"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "B"]  # Basic errors, flakes, bug risks

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"